More work on refactoring Excel Calculation Function Unit Tests
