+++ tests/PhpSpreadsheetTests/Calculation/Functions/LookupRef/HLookupTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\LookupRef;
+
+use PhpOffice\PhpSpreadsheet\Calculation\LookupRef;
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PHPUnit\Framework\TestCase;
+
+class HLookupTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerHLOOKUP
+     *
+     * @param mixed $expectedResult
+     */
+    public function testHLOOKUP($expectedResult, ...$args)
+    {
+        $result = LookupRef::HLOOKUP(...$args);
+        $this->assertEquals($expectedResult, $result);
+    }
+
+    public function providerHLOOKUP()
+    {
+        return require 'data/Calculation/LookupRef/HLOOKUP.php';
+    }
+}
