+++ tests/PhpSpreadsheetTests/Calculation/Functions/LookupRef/VLookupTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\LookupRef;
+
+use PhpOffice\PhpSpreadsheet\Calculation\LookupRef;
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PHPUnit\Framework\TestCase;
+
+class VLookupTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerVLOOKUP
+     *
+     * @param mixed $expectedResult
+     */
+    public function testVLOOKUP($expectedResult, ...$args)
+    {
+        $result = LookupRef::VLOOKUP(...$args);
+        $this->assertEquals($expectedResult, $result);
+    }
+
+    public function providerVLOOKUP()
+    {
+        return require 'data/Calculation/LookupRef/VLOOKUP.php';
+    }
+}
