+++ tests/PhpSpreadsheetTests/Calculation/LookupRefTest.php
@@ -20,54 +20,6 @@ class LookupRefTest extends TestCase
     }
 
     /**
-     * @dataProvider providerHLOOKUP
-     *
-     * @param mixed $expectedResult
-     */
-    public function testHLOOKUP($expectedResult, ...$args)
-    {
-        $result = LookupRef::HLOOKUP(...$args);
-        self::assertEquals($expectedResult, $result);
-    }
-
-    public function providerHLOOKUP()
-    {
-        return require 'data/Calculation/LookupRef/HLOOKUP.php';
-    }
-
-    /**
-     * @dataProvider providerVLOOKUP
-     *
-     * @param mixed $expectedResult
-     */
-    public function testVLOOKUP($expectedResult, ...$args)
-    {
-        $result = LookupRef::VLOOKUP(...$args);
-        self::assertEquals($expectedResult, $result);
-    }
-
-    public function providerVLOOKUP()
-    {
-        return require 'data/Calculation/LookupRef/VLOOKUP.php';
-    }
-
-    /**
-     * @dataProvider providerLOOKUP
-     *
-     * @param mixed $expectedResult
-     */
-    public function testLOOKUP($expectedResult, ...$args)
-    {
-        $result = LookupRef::LOOKUP(...$args);
-        self::assertEquals($expectedResult, $result);
-    }
-
-    public function providerLOOKUP()
-    {
-        return require 'data/Calculation/LookupRef/LOOKUP.php';
-    }
-
-    /**
      * @dataProvider providerMATCH
      *
      * @param mixed $expectedResult
